Graph echt ganz fertig mit Zeitlimitierung und Möglichkeit nur gespartes Co2 zu plotten, PoP-Up Messages implimentiert, Break Even Calc implimentiert wie auch ein paar andere kleinere Funktionen die gebraucht werden könnten.

+++ app/src/main/java/classes_and_Functions/VgFahrt.kt
@@ -74,17 +74,17 @@ fun createLineGraph(graph: GraphView, entries: List<VgFahrt>, seit: LocalDateTim
     )
     if(showCo2) graph.removeAllSeries()
 
-    var runningSum = 0
+    var runningSum = 0.0
 
     val series2 = LineGraphSeries(
 
         entries2.map { entries ->
 
             if(!showCo2)
-                runningSum += entries.price
+                runningSum += entries.price.toDouble()
 
             else
-                runningSum += entries.co2.toInt()
+                runningSum += entries.co2
 
             DataPoint(
